backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
twilio>=8.0.0
python-multipart>=0.0.9
supabase>=2.0.0
PyJWT>=2.8.0
